requirements.txt
docker==6.1.3
click==8.1.7
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0
